Bring default configuration for SRDataFile
This is required when we lookup the SRDataFile for files view controller
Basically we poll in the database for file objects. We get some of
them. But since they come from the DB, they don't get the SRCfg
object set.

TODO: for the proper configuration, I think I should add a setter
for the value that the singleton is returing.