Fix #1: don't an object from a wrong thread
Exporting is asynchronous and we pass and object there. This object can
potentially be used from a different thread, which Realm doesn't like. So
instead of object, I was an int representing the file ID. In the target
thread we just do the lookup again and get the object back.

For some reason it was less likely on a device. Easy to reproduce with the
simulator.